TC/NC & Removed old config stuff
Moved town/nation colors into their component
Removed save command that was apparently there since 2015
It still loads the town/nation colors from the old config if needed
Notification sound/pitch were the last config options to be converted over
#85
