+++ src/main/java/buttondevteam/chat/components/towncolors/admin/NationColorCommand.java
@@ -1,8 +1,10 @@
-package buttondevteam.chat.commands.ucmds.admin;
+package buttondevteam.chat.components.towncolors.admin;
 
 import buttondevteam.chat.PluginMain;
+import buttondevteam.chat.commands.ucmds.admin.AdminCommandBase;
+import buttondevteam.chat.components.towncolors.TownColorComponent;
+import buttondevteam.chat.components.towncolors.TownyListener;
 import buttondevteam.chat.components.towny.TownyComponent;
-import buttondevteam.chat.listener.TownyListener;
 import buttondevteam.lib.chat.Color;
 import com.palmergames.bukkit.towny.object.Nation;
 import com.palmergames.bukkit.towny.object.Town;
@@ -40,14 +42,14 @@ public class NationColorCommand extends AdminCommandBase {
 		val c = TownColorCommand.getColorOrSendError(args[1], sender);
 		if (!c.isPresent()) return true;
 		if (!c.get().getName().equals(Color.White.getName())) { //Default nation color
-			for (val nc : PluginMain.NationColor.values()) {
+			for (val nc : TownColorComponent.NationColor.values()) {
 				if (nc.getName().equals(c.get().getName())) {
 					sender.sendMessage("§cAnother nation already uses this color!");
 					return true;
 				}
 			}
 		}
-		PluginMain.NationColor.put(args[0].toLowerCase(), c.get());
+		TownColorComponent.NationColor.put(args[0].toLowerCase(), c.get());
 		Bukkit.getScheduler().runTaskAsynchronously(PluginMain.Instance, () -> {
 			for (Town t : nation.getTowns())
 				TownyListener.updateTownMembers(t);
